# api.yaml
openapi: 3.0.0
info:
  title: "Python CryptoPro API"
  version: "1.0"
servers:
  - 
    url: /api/{version}
    variables:
      version:
        description: API version
        default: v1
paths:
...
  /validate:
    post:
      operationId: api.sign.validate
      summary: Validate file signature
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - sign_type
                - file
              properties:
                sign_type:
                  description: Type of signature
                  type: string
                  enum: ['attached' , 'detached']
                file:
                  description: BASE64 encoded file content
                  type: string
                sign:
                  description: BASE64 encoded detached signature content
                  type: string
      responses:
        '200':
          description: Signature is valid
          content:
            application/json:
              schema:
                type: object
                properties:
                  valid:
                    type: boolean
        '400':
          description: Validation error
...